Move _get_reservoir_state_time_series to utils and rename to generate_reservoir_state_time_series
